Arithmetic operations with deb_lsd and deb_decimal

Implementation of arithmetic operations for pounds, shillings, and pence values as deb_lsd and deb_decimal vectors. Available operations are:

# S3 method for deb_lsd
vec_arith(op, x, y)

# S3 method for default
vec_arith.deb_lsd(op, x, y)

# S3 method for deb_lsd
vec_arith.deb_lsd(op, x, y)

# S3 method for numeric
vec_arith.deb_lsd(op, x, y)

# S3 method for deb_lsd
vec_arith.numeric(op, x, y)

# S3 method for MISSING
vec_arith.deb_lsd(op, x, y)

# S3 method for deb_decimal
vec_arith(op, x, y)

# S3 method for default
vec_arith.deb_decimal(op, x, y)

# S3 method for deb_decimal
vec_arith.deb_decimal(op, x, y)

# S3 method for numeric
vec_arith.deb_decimal(op, x, y)

# S3 method for deb_decimal
vec_arith.numeric(op, x, y)

# S3 method for MISSING
vec_arith.deb_decimal(op, x, y)

# S3 method for deb_decimal
vec_arith.deb_lsd(op, x, y)

# S3 method for deb_lsd
vec_arith.deb_decimal(op, x, y)

Arguments

op
Used internally to enable debkeepr to work with vctrs.

x, y
A pair of vectors.

Details

deb_lsd and deb_lsd: +, -, and /. The first two return a deb_lsd vector; the last returns a numeric vector.

deb_lsd and numeric: \* and /. Both return a deb_lsd vector.

numeric and deb_lsd: \*, returning a deb_lsd vector.

deb_decimal and deb_decimal: +, -, and /. The first two return a deb_lsd vector; the last returns a numeric vector.

deb_decimal and numeric: +, -, /, \*, ^, %%, %/%. All return a deb_decimal vector.

numeric and deb_decimal: +, -, \*. All return a deb_decimal vector.

deb_lsd and deb_decimal: +, -, /. The first two return a deb_lsd vector; the last returns a numeric vector.

deb_decimal and deb_lsd: +, -, /. The first two return a deb_lsd vector; the last returns a numeric vector.

Examples

# Arithmetic with deb_lsd
lsd1 <- deb_lsd(5, 16, 6)
lsd2 <- deb_lsd(7, 6, 8)

lsd1 + lsd2

#> <deb_lsd[1]>
#> [1] 13:3s:2d
#> # Bases: 20s 12d

lsd2 - lsd1

#> <deb_lsd[1]>
#> [1] 1:10s:2d
#> # Bases: 20s 12d

# Find the ration between two values
lsd2 / lsd1

#> [1] 1.258941

# Works with deb_lsd and numeric values
lsd2 / 2

#> <deb_lsd[1]>
#> [1] 3:13s:4d
#> # Bases: 20s 12d

2 * lsd1

#> <deb_lsd[1]>
#> [1] 11:13s:0d
#> # Bases: 20s 12d

# Arithmetic with deb_decimal
dec1 <- deb_decimal(8.45)
dec2 <- deb_decimal(4.625)

dec1 + dec2

#> <deb_decimal[1]>
#> [1] 13.075
#> # Unit: libra
#> # Bases: 20s 12d

dec1 - dec2

#> <deb_decimal[1]>
#> [1] 3.825
#> # Unit: libra
#> # Bases: 20s 12d

# Works with deb_decimal and numeric values
dec1 + dec2 + 7/3 + 8.25

#> <deb_decimal[1]>
#> [1] 23.65833
#> # Unit: libra
#> # Bases: 20s 12d

dec1 / 2

#> <deb_decimal[1]>
#> [1] 4.225
#> # Unit: libra
#> # Bases: 20s 12d

2 * dec2

#> <deb_decimal[1]>
#> [1] 9.25
#> # Unit: libra
#> # Bases: 20s 12d

# deb_lsd and deb_decimal returns a deb_lsd
# except with division

lsd1 + dec1

#> <deb_lsd[1]>
#> [1] 14:5s:6d
#> # Bases: 20s 12d

dec1 - lsd2

#> <deb_lsd[1]>
#> [1] 1:2s:4d
#> # Bases: 20s 12d

# When combing two deb-style vectors, the bases must match

# NOT RUN {
lsd1 + deb_lsd(5, 8, 3, bases = c(60, 16))
dec1 - deb_decimal(3.25, bases = c(60, 16))
# }

# When combining two `deb_decimal` vectors, the units must match

# NOT RUN {
dec1 + deb_decimal(23.25, unit = "s")
# }