Cast to deb_lsd

Cast x to a deb_lsd vector.

deb_as_lsd(x, ...)

# S3 method for deb_lsd
deb_as_lsd(x, ...)

# S3 method for deb_decimal
deb_as_lsd(x, ...)

# S3 method for numeric
deb_as_lsd(x, bases = c(20, 12), ...)

# S3 method for logical
deb_as_lsd(x, bases = c(20, 12), ...)

Arguments

x
An object to coerce to deb_lsd.

...
Arguments passed on to further methods.

bases
Numeric vector of length 2 used to specify the bases for the solidus or s and denarius or d units. Default is c(20, 12), which conforms to the most widely used system of 1 pound = 20 shillings and 1 shilling = 12 pence.

Value

A deb_lsd object.

Examples

# Cast a deb_decimal object to deb_lsd
x <- c(5.825, 3.25, 22/3)
d1 <- deb_decimal(x)
deb_as_lsd(d1)

#> <deb_lsd[3]>
#> [1] 5:16s:6d 3:5s:0d  7:6s:8d 
#> # Bases: 20s 12d

# Bases are automatically applied when
# casting from deb_decimal to deb_lsd
d2 <- deb_decimal(x, bases = c(60, 16))
deb_as_lsd(d2)

#> <deb_lsd[3]>
#> [1] 5:49s:8d 3:15s:0d 7:20s:0d
#> # Bases: 60s 16d

# Cast a numeric vector to deb_lsd
deb_as_lsd(x)

#> <deb_lsd[3]>
#> [1] 5:16s:6d 3:5s:0d  7:6s:8d 
#> # Bases: 20s 12d

# Use the bases argument to apply non-default bases
deb_as_lsd(x, bases = c(60, 16))

#> <deb_lsd[3]>
#> [1] 5:49s:8d 3:15s:0d 7:20s:0d
#> # Bases: 60s 16d