Accounts from the practice journal and ledger of Richard Dafforne

A data set of the accounts from the first practice journal and ledger in Richard Dafforne's accounting manual from 1660 called The Merchant's Mirrour. By 1660 The Merchant's Mirrour was in its third edition, and its contents had been printed in the well-known merchant manual of Gerard Malynes, Consuetudo Vel Lex Mercatoira since the 1636 edition, making it one of the most popular bookkeeping manuals in 17th-century England. The data set is meant to be used in conjunction with dafforne_transactions. It contains information on the accounts found in the practice journal and ledger that Dafforne used to teach double-entry bookkeeping practices.

dafforne_accounts

Format

A data frame with 46 rows and 5 variables.

Source

Richard Dafforne, The Merchant's Mirrour, Or Directions for the Perfect Ordering and Keeping of His Accounts, Third Edition, (London, 1660)

Details

The data set does not include the Balance account that Dafforne uses to close the books. The transactions from this account can be recreated using the lsd account functions in debkeepr.

Variables

id: Numeric id for each account. The ids correspond to the ids in the credit and debit variables in dafforne_transactions.

account: Name of the account.

ledger: Page on which the account appears in the ledger.

investor: The investor or the person’s whose capital is involved in the account. Accounts that only deal with the bookkeeper's capital are listed as "ego".

description: Short description of each account.