Calculate credit, debit, and current values of accounts

Calculate the total credit, debit, and the current values for all accounts in a transaction data frame with an lsd list column. Credits and debits are both returned as lsd list columns with positive values. If an account has more credit than debit, the current value will be returned as positive values. If the debit is greater, the current value will be returned as negative values.

deb_account_summary(df, credit = credit, debit = debit, lsd = lsd,
  round = 5, na.rm = FALSE)

Arguments

df
A data frame containing transactions between accounts and values in the form of an lsd list column.

credit
Credit column: Unquoted name of the credit variable. This represents the accounts that discharge the transactional values or from which the values derive. Default is credit. The credit column must be of the same class as the debit column.

debit
Debit column: Unquoted name of the debit variable. This represents the accounts that receive the transactional values. Default is debit. The debit column must be of the same class as the credit column.

lsd
lsd list column: Unquoted name of an lsd list column of pounds, shillings, and pence values with a bases attribute.

round
Round pence to specified number of decimal places. Default is 5. Rounding to whole numbers may give results that do not add together due to rounding rules.

na.rm
Logical. Passed on to na.rm argument in sum(). Whether missing values (NA) should be removed. Default is FALSE.

Value

Returns a tibble with one row for each account present in df and three lsd list columns representing the total credit, debit, and current values of the accounts.

Details

deb_account_summary() is similar to deb_account(), but it returns the values for all accounts in a transaction data frame instead of one.

deb_account_summary() is part of a family of functions meant to be used on data frames that contain transactions between accounts in an account book. The data frame should possess a similar structure to a directed edge list. In this context, credit and debit variables contain the ids of the accounts and act as the edges of the network. Following the nomenclature of accounting, the credit variable represents the account that the transactional value is from, while the debit variable represents the account that receives the value. Thus, from the credit account to the debit account.

See also

Examples

library(tibble)

# Create a transactions data frame
lsd_list <- deb_lsd(l = c(10, 6, 4, 7, 9),
                    s = c(15, 3, 11, 11, 2),
                    d = c(6, 11, 7, 8, 11),
                    bases = c(20, 12))
trans <- tibble(credit = c("a", "b", "b", "a", "c"),
                debit = c("b", "a", "a", "c", "a"),
                lsd = lsd_list)

# Credit, debit, and current values of accounts present in trans
deb_account_summary(df = trans,
                    credit = credit,
                    debit = debit,
                    lsd = lsd)

#> # A tibble: 3 x 4
#>   account_id credit    debit     current    
#>   <chr>      <lsd>     <lsd>     <lsd>      
#> 1 a          18, 7, 2  19, 18, 5 -1, -11, -3
#> 2 b          10, 15, 6 10, 15, 6 0, 0, 0    
#> 3 c          9, 2, 11  7, 11, 8  1, 11, 3