Cast deb_lsd to a list of lsd values

Cast a deb_lsd vector to a list of numeric vectors containing lsd values.

deb_as_list(x, ...)

Arguments

x
A deb_lsd vector to cast to a list of lsd values.

...
Arguments passed on to further methods.

Value

A list of numeric vectors of length 3, corresponding to lsd values.

Details

deb_as_list() turns a deb_lsd vector into a list of numeric vectors of length 3. It is the inverse of deb_as_lsd(). Compare to as.list(), which creates a list of deb_lsd vectors or unclass(), which creates a list of length 3 with numeric vectors for pounds, shillings, and pence.

See also

Examples

# deb_lsd vector
x <- deb_lsd(l = 0:3, s = 4:7, d = 8:11)

deb_as_list(x)

#> [[1]]
#> [1] 0 4 8
#> 
#> [[2]]
#> [1] 1 5 9
#> 
#> [[3]]
#> [1]  2  6 10
#> 
#> [[4]]
#> [1]  3  7 11
#> 

# This is the inverse of `deb_as_lsd()` of a list of lsd values
y <- deb_as_list(x)

identical(x, deb_as_lsd(y))

#> [1] TRUE