Exchange between pounds, shillings and pence currencies

Exchange between pounds, shillings, and pence currencies that share the same bases for shillings and pence given an exchange rate calculated on the basis of shillings.

deb_exchange(lsd, shillings_rate, bases = c(20, 12), round = 5)

Arguments

lsd
An lsd value. An object of class lsd or an object that can be coerced to class lsd: a numeric vector of length 3 or a list of such vectors.

shillings_rate
The exchange rate. This follows the common practice of calculating the exchange rate between different currencies in terms of shillings. Thus, in terms of a given number over the base of shillings provided in the bases argument. A numeric vector of length 1.

bases
Numeric vector of length 2 used to specify the bases for the shillings or s and pence or d units. Default is c(20, 12), which conforms to the most widely used system of 1 pound = 20 shillings and 1 shilling = 12 pence. If lsd is of class lsd, the bases attribute will be used in the place of this argument.

round
Round pence unit to specified number of decimal places. Default is 5. Set to 0 to return pence as whole number.

Value

Returns an object of class lsd with a bases attribute.

Details

deb_exchange() is a wrapper around deb_multiply(). It uses shillings as the basis for the conversion between the currencies. If the conversion between currencies is presented in pounds or pence, it may be easier to use deb_multiply(). If the bases for the shillings and pence units differ between the two currencies use deb_convert_bases().

Examples

# Exchange at the rate of 31 shillings
deb_exchange(lsd = c(850, 16, 10), shillings_rate = 31)

#>        l  s   d
#> [1] 1318 16 1.1

# Exchange between currencies that use the same alternative bases
deb_exchange(lsd = c(850, 16, 10),
             shillings_rate = 31,
             bases = c(20, 16))

#>        l  s    d
#> [1] 1318 15 12.3

# Exchange for an object of class lsd will use the bases attribute
lsd <- deb_as_lsd(lsd = c(850, 16, 10), bases = c(20, 16))
deb_exchange(lsd = lsd, shillings_rate = 31)

#>        l  s    d
#> [1] 1318 15 12.3

# If the exchange rate is in shillings and pence, you can either
# decimalize the shillings or add the pence and divide by 12 in
# the shillings_rate argument. If the decimalized shillings has
# a repeating decimal, the latter approach is preferable.

# Exchange at the rate of 31 shillings 4 pence
deb_exchange(lsd = c(850, 16, 10), shillings_rate = 31.33333)

#>        l  s       d
#> [1] 1332 19 8.43263

deb_exchange(lsd = c(850, 16, 10), shillings_rate = 31 + 4/12)

#>        l  s       d
#> [1] 1332 19 8.46667

# Exchange of a list of lsd values at a single rate
lsd_list <- list(c(40, 5, 9), c(29, 7, 1), c(35, 6, 5))
deb_exchange(lsd = lsd_list, shillings_rate = 31)

#>      l  s     d
#> [1] 62  8 10.95
#> [2] 45  9 11.75
#> [3] 54 14 11.35

# Or an lsd object with alternative bases
lsd_list2 <- deb_as_lsd(lsd = lsd_list, bases = c(20, 16))
deb_exchange(lsd = lsd_list2, shillings_rate = 31)

#>      l  s     d
#> [1] 62  8  9.95
#> [2] 45  9 15.15
#> [3] 54 14 12.55