Calculate the current values of open accounts

Calculate the current values of accounts in a transactions data frame and show only those accounts that have a positive or negative balance. If an account has more credit than debit, the current value will be returned as positive values. If the debit is greater, the current value will be returned as negative values.

deb_open(df, credit = credit, debit = debit, lsd = lsd, round = 5,
  na.rm = FALSE)

Arguments

df
A data frame containing transactions between accounts and values in the form of an lsd list column.

credit
Credit column: Unquoted name of the credit variable. This represents the accounts that discharge the transactional values or from which the values derive. Default is credit. The credit column must be of the same class as the debit column.

debit
Debit column: Unquoted name of the debit variable. This represents the accounts that receive the transactional values. Default is debit. The debit column must be of the same class as the credit column.

lsd
lsd list column: Unquoted name of an lsd list column of pounds, shillings, and pence values with a bases attribute.

round
Round pence to specified number of decimal places. Default is 5. Rounding to whole numbers may give results that do not add together due to rounding rules.

na.rm
Logical. Passed on to na.rm argument in sum(). Whether missing values (NA) should be removed. Default is FALSE.

Value

Returns a tibble with one row for each account in df where the current value in an lsd list column does not equal zero or does not have missing values in the current value. If all accounts are equal to zero or possess missing values, a tibble with zero rows will be returned.

Details

deb_open() is similar to deb_current(), but it only returns current values for accounts that have a value that is not zero.

deb_open() is part of a family of functions meant to be used on data frames that contain transactions between accounts in an account book. The data frame should possess a similar structure to a directed edge list. In this context, credit and debit variables contain the ids of the accounts and act as the edges of the network. Following the nomenclature of accounting, the credit variable represents the account that the transactional value is from, while the debit variable represents the account that receives the value. Thus, from the credit account to the debit account.

See also

Examples

library(tibble)

# Create a transactions data frame
lsd_list <- deb_lsd(l = c(10, 6, 4, 7, 9),
                    s = c(15, 3, 11, 11, 2),
                    d = c(6, 11, 7, 8, 11),
                    bases = c(20, 12))
trans <- tibble(credit = c("a", "b", "b", "a", "c"),
                debit = c("b", "a", "a", "c", "a"),
                lsd = lsd_list)

# Current values of open accounts present in trans
deb_open(df = trans,
         credit = credit,
         debit = debit,
         lsd = lsd)

#> [90m# A tibble: 2 x 2[39m
#>   account_id lsd        
#>   [3m[90m<chr>[39m[23m      [3m[90m<S3: lsd>[39m[23m  
#> [90m1[39m a          -1, -11, -3
#> [90m2[39m c          1, 11, 3