Replace NULL elements in an lsd object

Replace NULL elements in an lsd object with a numeric vector.

deb_replace_null(lsd, replace = c(0, 0, 0))

Arguments

lsd
An object of class lsd or a list of numeric vectors of length 3 that can be coerced to an lsd object.

replace
A numeric vector of length 3. Default is c(0, 0, 0), which replaces NULL with a value of 0.

Value

Returns an object of class lsd with NULL elements replaced.

Details

Note that NULL elements in an lsd object print as NA due to the printing method for lsd objects. NULL elements can be confirmed with str(). If c(NA, NA, NA) is used for replace argument, the NA vector will be coerced to numeric.

Examples

# Create lsd object with NULL elements
lsd <- deb_as_lsd(list(c(15, 4, 6),
                       NULL,
                       c(9, 12, 3),
                       NULL))
str(lsd)

#> List of 4
#>  $ : num [1:3] 15 4 6
#>  $ : NULL
#>  $ : num [1:3] 9 12 3
#>  $ : NULL
#>  - attr(*, "class")= chr "lsd"
#>  - attr(*, "bases")= num [1:2] 20 12

# Replace with default lsd value of 0
deb_replace_null(lsd = lsd)

#>      l  s d
#> [1] 15  4 6
#> [2]  0  0 0
#> [3]  9 12 3
#> [4]  0  0 0

# Replace NULL with NA
str(deb_replace_null(lsd = lsd, replace = c(NA, NA, NA)))

#> List of 4
#>  $ : num [1:3] 15 4 6
#>  $ : num [1:3] NA NA NA
#>  $ : num [1:3] 9 12 3
#>  $ : num [1:3] NA NA NA
#>  - attr(*, "class")= chr "lsd"
#>  - attr(*, "bases")= num [1:2] 20 12