Math group with deb_lsd vectors

Math and Summary group of functions with deb_lsd vectors. Implemented functions:

Summary group: sum(), any(), and all().

Math group: abs(), round(), signif(), ceiling(), floor(), trunc(), cummax(), cummin(), and cumsum().

Additional generics: mean(), is.nan(), is.finite(), and is.infinite().

All other functions from the groups not currently implemented, including median(), quantile(), and summary().

# S3 method for deb_lsd
sum(..., na.rm = FALSE)

# S3 method for deb_lsd
mean(x, ..., na.rm = FALSE)

# S3 method for deb_lsd
cumsum(x)

# S3 method for deb_lsd
round(x, digits = 0)

# S3 method for deb_lsd
signif(x, digits = 6)

# S3 method for deb_lsd
ceiling(x)

# S3 method for deb_lsd
floor(x)

# S3 method for deb_lsd
trunc(x, ...)

Arguments

...
deb_lsd vectors in sum() and arguments passed on to further methods in mean().

na.rm
Logical. Should missing values (including `NaN``) be removed?

x
An vector of class deb_lsd.

digits
Integer. Indicating the number of decimal places (round()) or significant digits (signif()) to be used.

Value

A deb_lsd vector with normalized values.

Details

sum() and cumsum() return a normalized deb_lsd values.

Round family of functions only affect the denarius (d) unit of a deb_lsd value. All values are normalized.

If you need a wider implementation of Math and Summary group functions, use a deb_decimal vector. However, median(), quantile(), and summary() are also not currently implemented for deb_decimal vectors. To use these functions cast deb_lsd and deb_decimal vectors to numeric.

Examples

x <- deb_lsd(l = c(5, 8, 12),
             s = c(16, 6, 13),
             d = c(6, 11, 0))

# All values are normalized with sum and cumsum
sum(x)

#> <deb_lsd[1]>
#> [1] 26:16s:5d
#> # Bases: 20s 12d

cumsum(x)

#> <deb_lsd[3]>
#> [1] 5:16s:6d  14:3s:5d  26:16s:5d
#> # Bases: 20s 12d

mean(x)

#> <deb_lsd[1]>
#> [1] 8:18s:9.667d
#> # Bases: 20s 12d

# Round family on deb_lsd affects the denarius unit
x2 <- deb_lsd(5, 12, 5.8365)
round(x2)

#> <deb_lsd[1]>
#> [1] 5:12s:6d
#> # Bases: 20s 12d

round(x2, digits = 2)

#> <deb_lsd[1]>
#> [1] 5:12s:5.84d
#> # Bases: 20s 12d

signif(x2, digits = 2)

#> <deb_lsd[1]>
#> [1] 5:12s:5.8d
#> # Bases: 20s 12d

ceiling(x2)

#> <deb_lsd[1]>
#> [1] 5:12s:6d
#> # Bases: 20s 12d

floor(x2)

#> <deb_lsd[1]>
#> [1] 5:12s:5d
#> # Bases: 20s 12d

trunc(x2)

#> <deb_lsd[1]>
#> [1] 5:12s:5d
#> # Bases: 20s 12d

# The returned values are normalized whether
# they are positive or negative
x3 <- deb_lsd(9, 19, 11.825)
x4 <- deb_lsd(-9, -19, -11.825)
round(x3)

#> <deb_lsd[1]>
#> [1] 10:0s:0d
#> # Bases: 20s 12d

round(x3, digits = 1)

#> <deb_lsd[1]>
#> [1] 9:19s:11.8d
#> # Bases: 20s 12d

ceiling(x3)

#> <deb_lsd[1]>
#> [1] 10:0s:0d
#> # Bases: 20s 12d

floor(x4)

#> <deb_lsd[1]>
#> [1] -10:0s:0d
#> # Bases: 20s 12d

trunc(x3)

#> <deb_lsd[1]>
#> [1] 9:19s:11d
#> # Bases: 20s 12d

trunc(x4)

#> <deb_lsd[1]>
#> [1] -9:-19s:-11d
#> # Bases: 20s 12d